// src/app/generated/api/model/collectionPermissionClaim.ts
/**
 * Gallery API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CollectionPermission } from './collectionPermission';


export interface CollectionPermissionClaim { 
    collectionId?: string;
    permissions?: Array<CollectionPermission> | null;
}